Applied extract method to layer filtering/ordering for geojson and blend renderers
