+++ lib/windshaft/renderers/blend/factory.js
@@ -4,6 +4,7 @@ var _ = require('underscore');
 
 var Renderer = require('./renderer');
 var BaseAdaptor = require('../base_adaptor');
+var layersFilter = require('../../utils/layer_filter');
 
 function BlendFactory(rendererFactory) {
     this.rendererFactory = rendererFactory;
@@ -36,21 +37,10 @@ BlendFactory.prototype.getRenderer = function(mapConfig, format, options, callba
     var filteredLayers = mapLayers.map(function(_layer, layerIdx) { return layerIdx; });
 
     if (layer !== 'all') {
-        filteredLayers = layer.split(',').map(function(layerIdx) {
-            return +layerIdx;
-        });
-
-        if (!filteredLayers.every(Number.isFinite)) {
-            return callback(new Error('Invalid layer filtering'));
-        }
-
-        filteredLayers = filteredLayers.sort(function(a, b) { return a - b; });
-
-        var uppermostLayerIdx = filteredLayers[filteredLayers.length - 1];
-        var lowestLayerIdx = filteredLayers[0];
-
-        if (lowestLayerIdx < 0 || uppermostLayerIdx >= mapLayers.length) {
-            return callback(new Error('Invalid layer filtering'));
+        try {
+            filteredLayers = layersFilter(mapConfig, mapLayers, layer);
+        } catch (err) {
+            return callback(err);
         }
     }
