+++ lib/windshaft/renderers/mapnik/factory.js
@@ -6,6 +6,7 @@ var assert = require('assert');
 var GeojsonRenderer = require('./geojson_renderer');
 var CartoPSQL = require('cartodb-psql');
 var RendererParams = require('../renderer_params');
+var layersFilter = require('../../utils/layer_filter');
 
 var MapnikAdaptor = require('./adaptor');
 var BaseAdaptor = require('../base_adaptor');
@@ -143,23 +144,12 @@ MapnikFactory.prototype.getGeoJsonRenderer = function(mapConfig, format, options
     if (typeof layerFilter === 'number') {
         layers = [ mapConfig.getLayer(layerFilter) ];
     } else if (layerFilter && layerFilter !== 'mapnik') {
-        var filteredLayers = layerFilter.split(',').map(function(layerIdx) {
-            return +layerIdx;
-        });
+        var filteredLayers;
 
-        if (!filteredLayers.every(Number.isFinite)) {
-            return callback(new Error('Invalid layer filtering'));
-        }
-
-        filteredLayers = filteredLayers.sort(function(a, b) {
-            return a - b;
-        });
-
-        var uppermostLayerIdx = filteredLayers[filteredLayers.length - 1];
-        var lowestLayerIdx = filteredLayers[0];
-
-        if (lowestLayerIdx < 0 || uppermostLayerIdx >= mapLayers.length) {
-            return callback(new Error('Invalid layer filtering'));
+        try {
+            filteredLayers = layersFilter(mapConfig, mapLayers, layerFilter);
+        } catch (err) {
+            return callback(err);
         }
 
         layers = filteredLayers.map(function (filteredLayer) {
